Move non api server related functionality to exe package
