+++ lib/exe/app/cardano-wallet.hs
@@ -20,7 +20,6 @@
 -- commands. Commands are turned into corresponding API calls, and submitted
 -- to an up-and-running server. Some commands do not require an active server
 -- and can be run "offline".
-
 module Main where
 
 import Prelude
@@ -58,10 +57,8 @@ import Cardano.Wallet.Api.Client
     , transactionClient
     , walletClient
     )
-import Cardano.Wallet.Api.Http.Shelley.Server
-    ( HostPreference
-    , Listen (..)
-    , TlsConfiguration
+import Cardano.Wallet.Api.Http.Server.Tls
+    ( TlsConfiguration
     )
 import Cardano.Wallet.Application
     ( TracerSeverities
@@ -105,6 +102,10 @@ import Cardano.Wallet.Application.CLI
     , tlsOption
     , tokenMetadataSourceOption
     , withLogging
+    )
+import Cardano.Wallet.Application.Server
+    ( HostPreference
+    , Listen
     )
 import Cardano.Wallet.Application.Version
     ( GitRevision
@@ -203,16 +204,18 @@ import qualified System.Info as I
 main :: IO ()
 main = withUtf8 $ do
     enableWindowsANSI
-    runCli $ cli $ mempty
-        <> cmdServe
-        <> cmdMnemonic
-        <> cmdKey
-        <> cmdWallet cmdWalletCreate walletClient
-        <> cmdAddress addressClient
-        <> cmdTransaction transactionClient walletClient
-        <> cmdNetwork networkClient
-        <> cmdStakePool stakePoolClient
-        <> cmdVersion
+    runCli
+        $ cli
+        $ mempty
+            <> cmdServe
+            <> cmdMnemonic
+            <> cmdKey
+            <> cmdWallet cmdWalletCreate walletClient
+            <> cmdAddress addressClient
+            <> cmdTransaction transactionClient walletClient
+            <> cmdNetwork networkClient
+            <> cmdStakePool stakePoolClient
+            <> cmdVersion
 
 beforeMainLoop :: Trace IO MainLog -> URI -> IO ()
 beforeMainLoop tr = logInfo tr . MsgListenAddress
@@ -234,73 +237,81 @@ data ServeArgs = ServeArgs
     , _poolMetadataSourceOpt :: Maybe PoolMetadataSource
     , _tokenMetadataSourceOpt :: Maybe TokenMetadataServer
     , _logging :: LoggingOptions TracerSeverities
-    } deriving (Show)
+    }
+    deriving (Show)
 
 cmdServe :: Mod CommandFields (IO ())
-cmdServe = command "serve" $ info (helper <*> helper' <*> cmd) $
-    progDesc "Serve API that listens for commands/actions."
+cmdServe =
+    command "serve"
+        $ info (helper <*> helper' <*> cmd)
+        $ progDesc "Serve API that listens for commands/actions."
   where
     helper' = helperTracing tracerDescriptions
 
-    cmd = fmap exec $ ServeArgs
-        <$> hostPreferenceOption
-        <*> modeOption nodeSocketOption
-        <*> listenApiOption
-        <*> listenUiOption
-        <*> optional tlsOption
-        <*> networkConfigurationOption
-        <*> optional databaseOption
-        <*> shutdownHandlerFlag
-        <*> optional poolMetadataSourceOption
-        <*> optional tokenMetadataSourceOption
-        <*> loggingOptions tracerSeveritiesOption
+    cmd =
+        fmap exec
+            $ ServeArgs
+                <$> hostPreferenceOption
+                <*> modeOption nodeSocketOption
+                <*> listenApiOption
+                <*> listenUiOption
+                <*> optional tlsOption
+                <*> networkConfigurationOption
+                <*> optional databaseOption
+                <*> shutdownHandlerFlag
+                <*> optional poolMetadataSourceOption
+                <*> optional tokenMetadataSourceOption
+                <*> loggingOptions tracerSeveritiesOption
 
     exec :: ServeArgs -> IO ()
-    exec args@(ServeArgs
-      host
-      mode
-      listenApi
-      listenUi
-      tlsConfig
-      networkConfig
-      databaseDir
-      enableShutdownHandler
-      poolMetadataFetching
-      tokenMetadataServerURI
-      logOpt) = withTracers logOpt $ \tr tracers -> do
-        withShutdownHandlerMaybe tr enableShutdownHandler $ do
-            logDebug tr $ MsgServeArgs args
+    exec
+        args@( ServeArgs
+                    host
+                    mode
+                    listenApi
+                    listenUi
+                    tlsConfig
+                    networkConfig
+                    databaseDir
+                    enableShutdownHandler
+                    poolMetadataFetching
+                    tokenMetadataServerURI
+                    logOpt
+                ) = withTracers logOpt $ \tr tracers -> do
+            withShutdownHandlerMaybe tr enableShutdownHandler $ do
+                logDebug tr $ MsgServeArgs args
 
-            (discriminant, netParams, vData, block0)
-                <- runExceptT (parseGenesisData networkConfig) >>= \case
+                (discriminant, netParams, vData, block0) <-
+                    runExceptT (parseGenesisData networkConfig) >>= \case
                         Right x -> pure x
                         Left err -> do
                             logError tr (MsgFailedToParseGenesis $ T.pack err)
                             exitWith $ ExitFailure 33
-            forM_ databaseDir $
-                setupDirectory (logInfo tr . MsgSetupDatabases)
+                forM_ databaseDir
+                    $ setupDirectory (logInfo tr . MsgSetupDatabases)
 
-            blockchainSource <- case mode of
-                Normal conn syncTolerance ->
-                    pure $ NodeSource conn vData syncTolerance
+                blockchainSource <- case mode of
+                    Normal conn syncTolerance ->
+                        pure $ NodeSource conn vData syncTolerance
 
-            exitWith =<< serveWallet
-                blockchainSource
-                netParams
-                tunedForMainnetPipeliningStrategy
-                discriminant
-                []
-                tracers
-                databaseDir
-                Nothing
-                host
-                listenApi
-                listenUi
-                tlsConfig
-                (Settings <$> poolMetadataFetching)
-                tokenMetadataServerURI
-                block0
-                (beforeMainLoop tr)
+                exitWith
+                    =<< serveWallet
+                        blockchainSource
+                        netParams
+                        tunedForMainnetPipeliningStrategy
+                        discriminant
+                        []
+                        tracers
+                        databaseDir
+                        Nothing
+                        host
+                        listenApi
+                        listenUi
+                        tlsConfig
+                        (Settings <$> poolMetadataFetching)
+                        tokenMetadataServerURI
+                        block0
+                        (beforeMainLoop tr)
 
     withShutdownHandlerMaybe :: Trace IO MainLog -> Bool -> IO () -> IO ()
     withShutdownHandlerMaybe _ False = void
@@ -328,10 +339,14 @@ data MainLog
 instance ToText MainLog where
     toText = \case
         MsgCmdLine exe args ->
-            T.pack $ unwords ("Command line:":exe:args)
+            T.pack $ unwords ("Command line:" : exe : args)
         MsgVersion ver rev arch os ->
-            "Running as " <> T.pack (showFullVersion ver rev) <> " on " <>
-            T.pack arch <> "-" <> T.pack os
+            "Running as "
+                <> T.pack (showFullVersion ver rev)
+                <> " on "
+                <> T.pack arch
+                <> "-"
+                <> T.pack os
         MsgSetupStateDir txt ->
             "Wallet state: " <> txt
         MsgSetupDatabases txt ->
@@ -346,14 +361,16 @@ instance ToText MainLog where
             "Interrupted by user."
         MsgShutdownHandler msg' ->
             toText msg'
-        MsgFailedToParseGenesis hint -> T.unwords
-            [ "Failed to parse Byron genesis configuration. You may want to check"
-            , "the filepath given via --genesis and make sure it points to a "
-            , "valid JSON Byron genesis file. The genesis file must be Byron, not"
-            , "Shelley as it used to feed the wallet with the initial blockchain"
-            , "parameters."
-            , "Here's (perhaps) some helpful hint:", hint
-            ]
+        MsgFailedToParseGenesis hint ->
+            T.unwords
+                [ "Failed to parse Byron genesis configuration. You may want to check"
+                , "the filepath given via --genesis and make sure it points to a "
+                , "valid JSON Byron genesis file. The genesis file must be Byron, not"
+                , "Shelley as it used to feed the wallet with the initial blockchain"
+                , "parameters."
+                , "Here's (perhaps) some helpful hint:"
+                , hint
+                ]
 
 withTracers
     :: LoggingOptions TracerSeverities
@@ -372,19 +389,22 @@ withTracers logOpt action =
         action trMain tracers `withException` logInterrupt
 
 tracerSeveritiesOption :: Parser TracerSeverities
-tracerSeveritiesOption = Tracers
-    <$> traceOpt applicationTracer (Just Info)
-    <*> traceOpt apiServerTracer (Just Info)
-    <*> traceOpt tokenMetadataTracer (Just Info)
-    <*> traceOpt walletEngineTracer (Just Info)
-    <*> traceOpt walletDbTracer (Just Info)
-    <*> traceOpt poolsEngineTracer (Just Info)
-    <*> traceOpt poolsDbTracer (Just Info)
-    <*> traceOpt ntpClientTracer (Just Info)
-    <*> traceOpt networkTracer (Just Info)
+tracerSeveritiesOption =
+    Tracers
+        <$> traceOpt applicationTracer (Just Info)
+        <*> traceOpt apiServerTracer (Just Info)
+        <*> traceOpt tokenMetadataTracer (Just Info)
+        <*> traceOpt walletEngineTracer (Just Info)
+        <*> traceOpt walletDbTracer (Just Info)
+        <*> traceOpt poolsEngineTracer (Just Info)
+        <*> traceOpt poolsDbTracer (Just Info)
+        <*> traceOpt ntpClientTracer (Just Info)
+        <*> traceOpt networkTracer (Just Info)
   where
-    traceOpt field def = fmap Const . option loggingSeverityOrOffReader $ mempty
-        <> long ("trace-" <> T.unpack (getConst (field tracerLabels)))
-        <> value def
-        <> metavar "SEVERITY"
-        <> internal
+    traceOpt field def =
+        fmap Const . option loggingSeverityOrOffReader
+            $ mempty
+                <> long ("trace-" <> T.unpack (getConst (field tracerLabels)))
+                <> value def
+                <> metavar "SEVERITY"
+                <> internal
